Replace GaussianLSTMPolicy
