+++ src/garage/tf/policies/__init__.py
@@ -23,8 +23,6 @@ from garage.tf.policies.gaussian_gru_policy import GaussianGRUPolicy
 from garage.tf.policies.gaussian_gru_policy_with_model import (
     GaussianGRUPolicyWithModel)
 from garage.tf.policies.gaussian_lstm_policy import GaussianLSTMPolicy
-from garage.tf.policies.gaussian_lstm_policy_with_model import (
-    GaussianLSTMPolicyWithModel)
 from garage.tf.policies.gaussian_mlp_policy import GaussianMLPPolicy
 
 __all__ = [
@@ -35,6 +33,5 @@ __all__ = [
     'CategoricalMLPPolicyWithModel', 'ContinuousMLPPolicy',
     'ContinuousMLPPolicyWithModel', 'DiscreteQfDerivedPolicy',
     'DeterministicMLPPolicy', 'GaussianGRUPolicy', 'GaussianLSTMPolicy',
-    'GaussianGRUPolicyWithModel', 'GaussianLSTMPolicyWithModel',
-    'GaussianMLPPolicy'
+    'GaussianGRUPolicyWithModel', 'GaussianMLPPolicy'
 ]
