+++ tests/garage/tf/models/test_gaussian_lstm_model.py
@@ -11,6 +11,7 @@ from tests.helpers import recurrent_step_lstm
 
 
 class TestGaussianLSTMModel(TfGraphTestCase):
+
     def setup_method(self):
         super().setup_method()
         self.batch_size = 1
@@ -22,8 +23,10 @@ class TestGaussianLSTMModel(TfGraphTestCase):
             (self.batch_size, self.time_step, self.feature_shape), 1.)
         self.obs_input = np.full((self.batch_size, self.feature_shape), 1.)
 
-        self.input_var = tf.compat.v1.placeholder(
-            tf.float32, shape=(None, None, self.feature_shape), name='input')
+        self.input_var = tf.compat.v1.placeholder(tf.float32,
+                                                  shape=(None, None,
+                                                         self.feature_shape),
+                                                  name='input')
         self.step_input_var = tf.compat.v1.placeholder(
             tf.float32, shape=(None, self.feature_shape), name='step_input')
 
@@ -38,25 +41,24 @@ class TestGaussianLSTMModel(TfGraphTestCase):
     def test_std_share_network_output_values(self, mock_normal, output_dim,
                                              hidden_dim):
         mock_normal.return_value = 0.5
-        model = GaussianLSTMModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=True,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        step_cell_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_cell',
-            dtype=tf.float32)
-        (action_var, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, step_cell, hidden_init_var, cell_init_var,
+        model = GaussianLSTMModel(output_dim=output_dim,
+                                  hidden_dim=hidden_dim,
+                                  std_share_network=True,
+                                  hidden_nonlinearity=None,
+                                  recurrent_nonlinearity=None,
+                                  hidden_w_init=self.default_initializer,
+                                  recurrent_w_init=self.default_initializer,
+                                  output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        step_cell_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                        hidden_dim),
+                                                 name='step_cell',
+                                                 dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         step_cell, hidden_init_var, cell_init_var,
          dist) = model.build(self.input_var, self.step_input_var,
                              step_hidden_var, step_cell_var)
 
@@ -70,27 +72,23 @@ class TestGaussianLSTMModel(TfGraphTestCase):
             feed_dict={self.input_var: self.obs_inputs})
 
         for i in range(self.time_step):
-            action, mean1, log_std1, hidden1, cell1 = self.sess.run(
-                [
-                    action_var, step_mean_var, step_log_std_var, step_hidden,
-                    step_cell
-                ],
+            mean1, log_std1, hidden1, cell1 = self.sess.run(
+                [step_mean_var, step_log_std_var, step_hidden, step_cell],
                 feed_dict={
                     self.step_input_var: self.obs_input,
                     step_hidden_var: hidden1,
                     step_cell_var: cell1
                 })
 
-            hidden2, cell2 = recurrent_step_lstm(
-                input_val=self.obs_input,
-                num_units=hidden_dim,
-                step_hidden=hidden2,
-                step_cell=cell2,
-                w_x_init=0.1,
-                w_h_init=0.1,
-                b_init=0.,
-                nonlinearity=None,
-                gate_nonlinearity=None)
+            hidden2, cell2 = recurrent_step_lstm(input_val=self.obs_input,
+                                                 num_units=hidden_dim,
+                                                 step_hidden=hidden2,
+                                                 step_cell=cell2,
+                                                 w_x_init=0.1,
+                                                 w_h_init=0.1,
+                                                 b_init=0.,
+                                                 nonlinearity=None,
+                                                 gate_nonlinearity=None)
 
             output_nonlinearity = np.full(
                 (np.prod(hidden2.shape[1:]), output_dim), 0.1)
@@ -100,9 +98,6 @@ class TestGaussianLSTMModel(TfGraphTestCase):
             assert np.allclose(hidden1, hidden2)
             assert np.allclose(cell1, cell2)
 
-            expected_action = 0.5 * np.exp(log_std1) + mean1
-            assert np.allclose(action, expected_action)
-
     # yapf: disable
     @pytest.mark.parametrize('output_dim, hidden_dim', [
         (1, 1),
@@ -111,25 +106,24 @@ class TestGaussianLSTMModel(TfGraphTestCase):
     ])
     # yapf: enable
     def test_std_share_network_shapes(self, output_dim, hidden_dim):
-        model = GaussianLSTMModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=True,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        step_cell_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_cell',
-            dtype=tf.float32)
-        (action_var, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, step_cell, hidden_init_var, cell_init_var,
+        model = GaussianLSTMModel(output_dim=output_dim,
+                                  hidden_dim=hidden_dim,
+                                  std_share_network=True,
+                                  hidden_nonlinearity=None,
+                                  recurrent_nonlinearity=None,
+                                  hidden_w_init=self.default_initializer,
+                                  recurrent_w_init=self.default_initializer,
+                                  output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        step_cell_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                        hidden_dim),
+                                                 name='step_cell',
+                                                 dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         step_cell, hidden_init_var, cell_init_var,
          dist) = model.build(self.input_var, self.step_input_var,
                              step_hidden_var, step_cell_var)
 
@@ -151,29 +145,29 @@ class TestGaussianLSTMModel(TfGraphTestCase):
                                                                   (3, 3, 1),
                                                                   (3, 3, 2)])
     @mock.patch('tensorflow.random.normal')
-    def test_without_std_share_network_output_values(
-            self, mock_normal, output_dim, hidden_dim, init_std):
+    def test_without_std_share_network_output_values(self, mock_normal,
+                                                     output_dim, hidden_dim,
+                                                     init_std):
         mock_normal.return_value = 0.5
-        model = GaussianLSTMModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=False,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer,
-            init_std=init_std)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        step_cell_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_cell',
-            dtype=tf.float32)
-        (action_var, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, step_cell, hidden_init_var, cell_init_var,
+        model = GaussianLSTMModel(output_dim=output_dim,
+                                  hidden_dim=hidden_dim,
+                                  std_share_network=False,
+                                  hidden_nonlinearity=None,
+                                  recurrent_nonlinearity=None,
+                                  hidden_w_init=self.default_initializer,
+                                  recurrent_w_init=self.default_initializer,
+                                  output_w_init=self.default_initializer,
+                                  init_std=init_std)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        step_cell_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                        hidden_dim),
+                                                 name='step_cell',
+                                                 dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         step_cell, hidden_init_var, cell_init_var,
          dist) = model.build(self.input_var, self.step_input_var,
                              step_hidden_var, step_cell_var)
 
@@ -187,27 +181,23 @@ class TestGaussianLSTMModel(TfGraphTestCase):
             feed_dict={self.input_var: self.obs_inputs})
 
         for i in range(self.time_step):
-            action, mean1, log_std1, hidden1, cell1 = self.sess.run(
-                [
-                    action_var, step_mean_var, step_log_std_var, step_hidden,
-                    step_cell
-                ],
+            mean1, log_std1, hidden1, cell1 = self.sess.run(
+                [step_mean_var, step_log_std_var, step_hidden, step_cell],
                 feed_dict={
                     self.step_input_var: self.obs_input,
                     step_hidden_var: hidden1,
                     step_cell_var: cell1
                 })
 
-            hidden2, cell2 = recurrent_step_lstm(
-                input_val=self.obs_input,
-                num_units=hidden_dim,
-                step_hidden=hidden2,
-                step_cell=cell2,
-                w_x_init=0.1,
-                w_h_init=0.1,
-                b_init=0.,
-                nonlinearity=None,
-                gate_nonlinearity=None)
+            hidden2, cell2 = recurrent_step_lstm(input_val=self.obs_input,
+                                                 num_units=hidden_dim,
+                                                 step_hidden=hidden2,
+                                                 step_cell=cell2,
+                                                 w_x_init=0.1,
+                                                 w_h_init=0.1,
+                                                 b_init=0.,
+                                                 nonlinearity=None,
+                                                 gate_nonlinearity=None)
 
             output_nonlinearity = np.full(
                 (np.prod(hidden2.shape[1:]), output_dim), 0.1)
@@ -219,9 +209,6 @@ class TestGaussianLSTMModel(TfGraphTestCase):
             assert np.allclose(hidden1, hidden2)
             assert np.allclose(cell1, cell2)
 
-            expected_action = 0.5 * np.exp(log_std1) + mean1
-            assert np.allclose(action, expected_action)
-
     # yapf: disable
     @pytest.mark.parametrize('output_dim, hidden_dim', [
         (1, 1),
@@ -230,25 +217,24 @@ class TestGaussianLSTMModel(TfGraphTestCase):
     ])
     # yapf: enable
     def test_without_std_share_network_shapes(self, output_dim, hidden_dim):
-        model = GaussianLSTMModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=False,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        step_cell_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_cell',
-            dtype=tf.float32)
-        (action_var, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, step_cell, hidden_init_var, cell_init_var,
+        model = GaussianLSTMModel(output_dim=output_dim,
+                                  hidden_dim=hidden_dim,
+                                  std_share_network=False,
+                                  hidden_nonlinearity=None,
+                                  recurrent_nonlinearity=None,
+                                  hidden_w_init=self.default_initializer,
+                                  recurrent_w_init=self.default_initializer,
+                                  output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        step_cell_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                        hidden_dim),
+                                                 name='step_cell',
+                                                 dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         step_cell, hidden_init_var, cell_init_var,
          dist) = model.build(self.input_var, self.step_input_var,
                              step_hidden_var, step_cell_var)
 
@@ -277,27 +263,25 @@ class TestGaussianLSTMModel(TfGraphTestCase):
     def test_std_share_network_is_pickleable(self, mock_normal, output_dim,
                                              hidden_dim):
         mock_normal.return_value = 0.5
-        model = GaussianLSTMModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=True,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        step_cell_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_cell',
-            dtype=tf.float32)
-        (_, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, step_cell, _, _, _) = model.build(
-             self.input_var, self.step_input_var, step_hidden_var,
-             step_cell_var)
+        model = GaussianLSTMModel(output_dim=output_dim,
+                                  hidden_dim=hidden_dim,
+                                  std_share_network=True,
+                                  hidden_nonlinearity=None,
+                                  recurrent_nonlinearity=None,
+                                  hidden_w_init=self.default_initializer,
+                                  recurrent_w_init=self.default_initializer,
+                                  output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        step_cell_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                        hidden_dim),
+                                                 name='step_cell',
+                                                 dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         step_cell, _, _, _) = model.build(self.input_var, self.step_input_var,
+                                           step_hidden_var, step_cell_var)
 
         # output layer is a tf.keras.layers.Dense object,
         # which cannot be access by tf.compat.v1.variable_scope.
@@ -323,26 +307,27 @@ class TestGaussianLSTMModel(TfGraphTestCase):
         with tf.compat.v1.Session(graph=tf.Graph()) as sess:
             model_pickled = pickle.loads(h)
 
-            input_var = tf.compat.v1.placeholder(
-                tf.float32,
-                shape=(None, None, self.feature_shape),
-                name='input')
+            input_var = tf.compat.v1.placeholder(tf.float32,
+                                                 shape=(None, None,
+                                                        self.feature_shape),
+                                                 name='input')
             step_input_var = tf.compat.v1.placeholder(
                 tf.float32,
                 shape=(None, self.feature_shape),
                 name='step_input')
-            step_hidden_var = tf.compat.v1.placeholder(
-                shape=(self.batch_size, hidden_dim),
-                name='initial_hidden',
-                dtype=tf.float32)
-            step_cell_var = tf.compat.v1.placeholder(
-                shape=(self.batch_size, hidden_dim),
-                name='initial_cell',
-                dtype=tf.float32)
+            step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                              hidden_dim),
+                                                       name='initial_hidden',
+                                                       dtype=tf.float32)
+            step_cell_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                            hidden_dim),
+                                                     name='initial_cell',
+                                                     dtype=tf.float32)
 
-            (_, mean_var2, step_mean_var2, log_std_var2, step_log_std_var2,
-             step_hidden2, step_cell2, _, _, _) = model_pickled.build(
-                 input_var, step_input_var, step_hidden_var, step_cell_var)
+            (mean_var2, step_mean_var2, log_std_var2, step_log_std_var2,
+             step_hidden2, step_cell2, _, _,
+             _) = model_pickled.build(input_var, step_input_var,
+                                      step_hidden_var, step_cell_var)
 
             outputs2 = sess.run([mean_var2, log_std_var2],
                                 feed_dict={input_var: self.obs_inputs})
@@ -367,27 +352,25 @@ class TestGaussianLSTMModel(TfGraphTestCase):
     def test_without_std_share_network_is_pickleable(self, mock_normal,
                                                      output_dim, hidden_dim):
         mock_normal.return_value = 0.5
-        model = GaussianLSTMModel(
-            output_dim=output_dim,
-            hidden_dim=hidden_dim,
-            std_share_network=False,
-            hidden_nonlinearity=None,
-            recurrent_nonlinearity=None,
-            hidden_w_init=self.default_initializer,
-            recurrent_w_init=self.default_initializer,
-            output_w_init=self.default_initializer)
-        step_hidden_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_hidden',
-            dtype=tf.float32)
-        step_cell_var = tf.compat.v1.placeholder(
-            shape=(self.batch_size, hidden_dim),
-            name='step_cell',
-            dtype=tf.float32)
-        (_, mean_var, step_mean_var, log_std_var, step_log_std_var,
-         step_hidden, step_cell, _, _, _) = model.build(
-             self.input_var, self.step_input_var, step_hidden_var,
-             step_cell_var)
+        model = GaussianLSTMModel(output_dim=output_dim,
+                                  hidden_dim=hidden_dim,
+                                  std_share_network=False,
+                                  hidden_nonlinearity=None,
+                                  recurrent_nonlinearity=None,
+                                  hidden_w_init=self.default_initializer,
+                                  recurrent_w_init=self.default_initializer,
+                                  output_w_init=self.default_initializer)
+        step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                          hidden_dim),
+                                                   name='step_hidden',
+                                                   dtype=tf.float32)
+        step_cell_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                        hidden_dim),
+                                                 name='step_cell',
+                                                 dtype=tf.float32)
+        (mean_var, step_mean_var, log_std_var, step_log_std_var, step_hidden,
+         step_cell, _, _, _) = model.build(self.input_var, self.step_input_var,
+                                           step_hidden_var, step_cell_var)
 
         # output layer is a tf.keras.layers.Dense object,
         # which cannot be access by tf.compat.v1.variable_scope.
@@ -413,26 +396,27 @@ class TestGaussianLSTMModel(TfGraphTestCase):
         with tf.compat.v1.Session(graph=tf.Graph()) as sess:
             model_pickled = pickle.loads(h)
 
-            input_var = tf.compat.v1.placeholder(
-                tf.float32,
-                shape=(None, None, self.feature_shape),
-                name='input')
+            input_var = tf.compat.v1.placeholder(tf.float32,
+                                                 shape=(None, None,
+                                                        self.feature_shape),
+                                                 name='input')
             step_input_var = tf.compat.v1.placeholder(
                 tf.float32,
                 shape=(None, self.feature_shape),
                 name='step_input')
-            step_hidden_var = tf.compat.v1.placeholder(
-                shape=(self.batch_size, hidden_dim),
-                name='initial_hidden',
-                dtype=tf.float32)
-            step_cell_var = tf.compat.v1.placeholder(
-                shape=(self.batch_size, hidden_dim),
-                name='initial_cell',
-                dtype=tf.float32)
+            step_hidden_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                              hidden_dim),
+                                                       name='initial_hidden',
+                                                       dtype=tf.float32)
+            step_cell_var = tf.compat.v1.placeholder(shape=(self.batch_size,
+                                                            hidden_dim),
+                                                     name='initial_cell',
+                                                     dtype=tf.float32)
 
-            (_, mean_var2, step_mean_var2, log_std_var2, step_log_std_var2,
-             step_hidden2, step_cell2, _, _, _) = model_pickled.build(
-                 input_var, step_input_var, step_hidden_var, step_cell_var)
+            (mean_var2, step_mean_var2, log_std_var2, step_log_std_var2,
+             step_hidden2, step_cell2, _, _,
+             _) = model_pickled.build(input_var, step_input_var,
+                                      step_hidden_var, step_cell_var)
 
             outputs2 = sess.run([mean_var2, log_std_var2],
                                 feed_dict={input_var: self.obs_inputs})
